working on edit topic

+++ src/User/index.js
@@ -7,10 +7,7 @@ class User extends Component {
         super(props);
 
         this.state = {
-            username:   '',
-            email:      '',
-            diplayName: '',
-            topics:     []            
+            topics: []
         }
     }
 
@@ -20,7 +17,7 @@ class User extends Component {
 
     getUserTopics = async () => {
         try {
-            const response = await fetch('http://localhost:9000/users/' + this.props.userInfo.userId, {
+            const response = await fetch('http://localhost:9000/users/' + this.props.loggedInUserId, {
                 method:      'GET',
                 credentials: 'include'
             });
@@ -30,13 +27,12 @@ class User extends Component {
             }
 
             const userParsed = await response.json();
-            
+
+            this.props.handleCompleteUserInfo(userParsed);
+            this.props.handleCompleteUserInfo(userParsed.userTopics);
             this.setState({
-                username:    userParsed.user.username,
-                email:       userParsed.user.email,
-                displayName: userParsed.user.displayName,
-                topics:      userParsed.user.topics
-            })            
+                topics:      userParsed.userTopics
+            })
         } catch (err) {
             console.log(err);
             return err;
@@ -45,15 +41,21 @@ class User extends Component {
 
     deleteUser = async (e) => {
         try {
-            const deleteUser = await fetch('http://localhost:9000/users/' + this.props.userInfo.userId, {
-               method:  'DELETE',
-               credentials: 'include' 
-            });
-            const parsedResponse = await deleteUser.json();
+            let verify = window.confirm('Are you sure!!');
 
-            if(parsedResponse.status === 200) {
-                alert('Sorry! You Delete your profile.');
-                this.props.history.push('/');
+            if(verify) {
+                const deleteUser = await fetch('http://localhost:9000/users/' + this.props.loggedInUserId, {
+                    method:  'DELETE',
+                    credentials: 'include' 
+                 });
+                 const parsedResponse = await deleteUser.json();
+     
+                 if(parsedResponse.status === 200) {
+                     alert('Sorry! You Delete your profile.');
+                     this.props.history.push('/');
+                 }
+            } else {
+                alert('Good Choice!');
             }
         } catch (err) {
             console.log(err);
@@ -61,22 +63,24 @@ class User extends Component {
         }
     }
 
-    render() {      
+    render() {  
+        console.log(this.state.topics, 'in user state') 
         return(
             <div>
                 <label>
-                    <h2>Welcome: {this.state.username}</h2>
+                    <h2>Welcome: {this.props.loggedInUsername}</h2>
                 </label><br/>
                 <label>
-                    <button name = "edit" onClick = {() => this.props.history.push({
-                                                     pathname: '/user/edit', 
-                                                     state: {userInfo: this.state}
-                                                     })
-                                                    } > Edit Your Profile</button><br/>
-                    <button name = "delete" onClick = {this.deleteUser} >Delete Your Profile</button><br/>
+                    <button name = 'edit' onClick = {() => this.props.history.push('/user/edit')} >Edit Your Profile</button><br/>
+                </label>
+                <label>
+                    <button name = 'delete' onClick = {this.deleteUser} >Delete Your Profile</button><br/>
                 </label>
                 <label>
                     <h3>All Topics: <TopicsList topics = {this.state.topics} /></h3>
+                </label>
+                <label>
+                    <h3><button onClick = {() => this.props.history.push('/user/newtopic')}>Add new Topics</button></h3>
                 </label>
             </div>
         )
